Big jump, update to August 24th, 2024 upstream commit.
Builds fine without GVFS, the option is in the Makefile but commented out.  This has not been tested to see if it operates properly, but a successful build with an optional capability left out is darn good.  I assume that things are installed into their proper locations for python, etc. according to the FreeBSD ports framework automation.